Finish up to dashboard

+++ server/schemas/resolvers.js
@@ -7,24 +7,43 @@ const resolvers = {
         tags: async () => {
             return await Tag.find();
         },
-        user: async (parent, args, context) => {
+        me: async (parent, args, context) => {
             if (context.user) {
                 const userData = await User.findById(context.user._id)
                     .select('-__v -password')
-                    .populate('activities')
-                    .populate({ 
+                    .populate({
+                        path: 'activities',
+                        populate: {
+                            path: 'tags author comments likes'
+                        }
+                    })
+                    .populate({
                         path: 'comments',
                         populate: {
-                          path: 'activity'
-                        } 
+                            path: 'activity'
+                        }
                     })
-                    .populate({ 
+                    .populate({
                         path: 'likes',
                         populate: {
-                          path: 'activity'
-                        } 
+                            path: 'activity'
+                        }
                     })
 
+                return userData
+            }
+            throw new AuthenticationError('Not logged in');
+        },
+        user: async (parent, { _id }, context) => {
+            if (context.user) {
+                const userData = await User.findById(_id)
+                    .select('-__v -password')
+                    .populate({
+                        path: 'activities',
+                        populate: {
+                            path: 'tags'
+                        }
+                    });
                 return userData
             }
             throw new AuthenticationError('Not logged in');
@@ -40,21 +59,23 @@ const resolvers = {
                     $regex: title
                 };
             }
-            
+
             return await Activity.find(params)
-                            .populate('author')
-                            .populate('tags');
+                .populate('author')
+                .populate('tags')
+                .sort({ createdAt: -1 })
         },
         activity: async (parent, { _id }) => {
             return await Activity.findById(_id)
-                            .populate('author')
-                            .populate('tags')
-                            .populate({ 
-                                path: 'comments',
-                                populate: {
-                                  path: 'user'
-                                } 
-                            })
+                .populate('author')
+                .populate('tags')
+                .populate('likes')
+                .populate({
+                    path: 'comments',
+                    populate: {
+                        path: 'user'
+                    }
+                })
         }
     },
     Mutation: {
@@ -66,11 +87,11 @@ const resolvers = {
         },
         addActivity: async (parent, args, context) => {
             if (context.user) {
-                const activity = await Activity.create(args);
-                
+                const activity = await Activity.create({ ...args, author: context.user._id });
+
                 await User.findByIdAndUpdate(
-                    {_id: context.user._id}, 
-                    { $push: { activities: activity._id }},
+                    { _id: context.user._id },
+                    { $push: { activities: activity._id } },
                     { new: true }
                 );
                 return activity;
@@ -79,15 +100,15 @@ const resolvers = {
         },
         addComment: async (parent, args, context) => {
             if (context.user) {
-                const comment = await Comment.create(args);
+                const comment = await (await Comment.create({ ...args, user: context.user._id })).populate("user");
 
                 await User.findByIdAndUpdate(
-                    { _id: context.user._id}, 
-                    { $push: { comments: comment._id }},
+                    { _id: context.user._id },
+                    { $push: { comments: comment._id } },
                     { new: true }
                 );
-    
-                const activity = await Activity.findByIdAndUpdate(
+
+                await Activity.findByIdAndUpdate(
                     { _id: args.activity },
                     { $push: { comments: comment._id } },
                     { new: true }
@@ -97,13 +118,13 @@ const resolvers = {
             }
             throw new AuthenticationError('Not logged in!');
         },
-        addLike: async ( parent, args, context ) => {
+        addLike: async (parent, args, context) => {
             if (context.user) {
-                const like = await Like.create(args);
+                const like = await Like.create({ ...args, user: context.user._id });
 
                 await User.findByIdAndUpdate(
-                    {_id: context.user._id}, 
-                    { $push: { likes: like._id }},
+                    { _id: context.user._id },
+                    { $push: { likes: like._id } },
                     { new: true }
                 );
 
@@ -119,7 +140,7 @@ const resolvers = {
         },
         updateUser: async (parent, args, context) => {
             if (context.user) {
-                return await User.findByIdAndUpdate(context.user._id, args, { new: true});
+                return await User.findByIdAndUpdate(context.user._id, args, { new: true });
             }
 
             throw new AuthenticationError('Not logged in!');
@@ -127,15 +148,27 @@ const resolvers = {
         updateActivity: async (parent, args, context) => {
             if (context.user) {
                 const { _id, ...updateContent } = args;
-                return await Activity.findByIdAndUpdate(_id, updateContent, { new: true });
+                const activity = await Activity.findById(_id);
+
+                if (activity.author._id.equals(context.user._id)) {
+                    return await Activity.findByIdAndUpdate(_id, updateContent, { new: true });
+                };
+
+                throw new AuthenticationError('You have no authentication to update this activity.')
             }
 
             throw new AuthenticationError('Not logged in!')
         },
         updateComment: async (parent, args, context) => {
             if (context.user) {
-                const { _id, ...updateContent} = args;
-                return await Comment.findByIdAndUpdate(_id, updateContent, { new: true });
+                const { _id, ...updateContent } = args;
+                const comment = await Comment.findById(_id);
+
+                if (comment.user._id.equals(context.user._id)) {
+                    return await Comment.findByIdAndUpdate(_id, updateContent, { new: true });
+                };
+
+                throw new AuthenticationError('You have no authentication to update this comment.')
             }
             throw new AuthenticationError('Not logged in!')
         },
@@ -156,36 +189,65 @@ const resolvers = {
 
             return { token, user };
         },
-        deleteActivity: async (parent, {_id}, context) => {
+        deleteActivity: async (parent, { _id }, context) => {
             if (context.user) {
-                await Activity.findByIdAndDelete(_id);
-                return await User.findByIdAndUpdate(
-                    {_id: context.user._id},
-                    { $pull: { activities: _id }},
-                    { new: true }
-                )
+                const activity = await Activity.findById(_id);
+
+                if (activity.author._id.equals(context.user._id)) {
+                    await Activity.findByIdAndDelete(_id);
+                    return await User.findByIdAndUpdate(
+                        { _id: context.user._id },
+                        { $pull: { activities: { _id: _id } } },
+                        { new: true }
+                    )
+                }
+
+                throw new AuthenticationError('You have no authentication to delete this activity.')
+
             }
             throw new AuthenticationError('Not logged in!');
         },
-        deleteComment: async (parent, {_id}, context) => {
+        deleteComment: async (parent, { _id }, context) => {
             if (context.user) {
-                await Comment.findByIdAndDelete(_id);
-                return await User.findByIdAndUpdate(
-                    {_id: context.user._id},
-                    { $pull: { comments: _id }},
-                    { new: true }
-                )
+                const comment = await Comment.findById(_id);
+
+                if (comment.user.equals(context.user._id)) {
+                    await User.findByIdAndUpdate(
+                        { _id: context.user._id },
+                        { $pull: { comments: { _id: _id } } },
+                        { new: true }
+                    )
+                    await Activity.findByIdAndUpdate(
+                        { _id: comment.activity },
+                        { $pull: { comments: { _id: _id } } },
+                        { new: true }
+
+                    )
+                    return await Comment.findByIdAndDelete(_id);
+                }
+                throw new AuthenticationError('You have no authentication to delete this comment.')
             }
             throw new AuthenticationError('Not logged in!');
         },
-        deleteLike: async (parent, {_id}, context) => {
+        deleteLike: async (parent, { _id }, context) => {
             if (context.user) {
-                await Like.findByIdAndDelete(_id);
-                return await User.findByIdAndDelete(
-                    {_id: context.user._id},
-                    { $pull: { likes: _id }},
-                    { new: true }
-                )
+                const like = await Like.findById(_id);
+                console.log(like);
+
+                if (like.user.equals(context.user._id)) {
+                    await User.findByIdAndUpdate(
+                        { _id: context.user._id },
+                        { $pull: { likes: { _id: _id } } },
+                        { new: true }
+                    );
+                    await Activity.findByIdAndUpdate(
+                        { _id: like.activity._id },
+                        { $pull: { likes: { _id: _id } } },
+                        { new: true }
+                    );
+                    return await Like.findByIdAndDelete(_id);
+                };
+                throw new AuthenticationError('You have no authentication to undo the like.')
             }
             throw new AuthenticationError('Not logged in!');
         }
